Refactoring and rearranging functions
